Slides/requirements.txt
python-pptx==1.0.2
pandas>=2.2.3 